exportHTML {iNZightPlots}	R Documentation

ExportHTML

Description

exportHTML is designed to export the iNZight plot as a dynamic, interactive HTML page. Currently only handles single panel plots. Coloured hex plots are currently not available yet.

Usage

exportHTML(x, file, data, extra.vars, ...)

## S3 method for class 'function'
exportHTML(x, file = "index.html", data = NULL,
  extra.vars = NULL, width = dev.size()[1], height = dev.size()[2], ...)

## S3 method for class 'ggplot'
exportHTML(x, file = "index.html", data = NULL,
  extra.vars = NULL, mapObj, ...)

## S3 method for class 'inzplotoutput'
exportHTML(x, file = "index.html", data = NULL,
  extra.vars = NULL, ...)

Arguments

x
An iNZight plot object that captures iNZight environment

file
Name of temporary HTML file generated Additional parameters for scatterplots and dotplots only:

data
dataset/dataframe that you wish to investigate and export more variables from

extra.vars
extra variables specified by the user to be exported

...
extra arguments

width
the desired width of the SVG plot

height
the desired height of the SVG plot

mapObj
iNZightMap object (from iNZightMaps)

Value

Opens up an HTML file of x with filename file in the browser (best performance on Chrome/Firefox)

Methods (by class)

function: method for an iNZightPlot-generating function

ggplot: method for iNZightMaps output (via new module)

inzplotoutput: method for output from iNZightPlot

Author(s)

Yu Han Soh

Examples

## Not run: 
x <- iNZightPlot(Petal.Width, Petal.Length, data = iris, colby = Species)
exportHTML(x, "index.html")

#to export more variables for scatterplots:
 exportHTML(x, "index.html", data = iris, extra.vars = c("Sepal.Length", "Sepal.Width"))

## End(Not run)

[Package iNZightPlots version 2.7.5 Index]