exportSVG {iNZightPlots}	R Documentation

Export iNZightPlots as an SVG

Description

exportSVG is designed to export the iNZight plot as a temporary SVG that is opened in a web browser. The iNZightPlot must be drawn to a graphics device before exporting can occur.

Usage

exportSVG(x, file)

Arguments

x
iNZight plot object or function that captures iNZight environment

file
Name of temporary svg file generated (by default: 'inzightplot.svg')

Value

Opens up an SVG file of x with filename file in a web browser

Author(s)

Yu Han Soh

[Package iNZightPlots version 2.7.5 Index]