inzpar {iNZightPlots}	R Documentation

iNZight Plotting Parameters

Description

Plotting parameters for iNZight Plots

Usage

inzpar(..., .viridis = requireNamespace("viridis", quietly = TRUE))

Arguments

...
If arguments are supplied, then these values are set. If left empty, then

.viridis
checks if the viridis package is installed; or can be turend off the default list is returned.

Details

A whole suite of parameters that can be used to fine-tune plots obtained from the iNZightPlot function. The parameters include both plot type, style, and appearance. They are described below.

'pch'

the plotting symbol to be used; default is '1' (empty circle)

'col.pt'

the colour of points. this can either be a single value, or a vector of colours if colby is specified

'col.missing'

the colour for missing values; default is a light grey

'cex'

the overal scaling for the entire plot; values less than 1 will make the text and points smaller, while values larger than 1 will magnify everything

'cex.pt'

the scaling value for points

'cex.dotpt'

the scaling value for points in a dotplot. Note, this is not multiplicative with 'cex.pt'

'cex.lab'

the scaling value for the plot labels

'cex.axis'

the scaling value for the axis labels

'cex.main'

the scaling value for the main plot title

'cex.text'

the scaling value for text on the plot

'alpha'

transparency setting for points; default is 1, 0 is fully transparent

'bg'

the background color for the plot

'fill.pt'

the fill colour for points; default is "transparent"

'lwd'

the line width of lines (for joining points)

'lty'

the line type of lines (for joining points)

'lwd.pt'

the line width used for points; default is 2

'col.line'

the colour of lines used to join points

'col.sub'

vector of up to two colours for the background of subplot labels. If only one specified, it is used for both.

'locate.col.def'

the default colour for locating points

'jitter'

the axes to add jitter to. Takes values "x", "y", or "xy" (default is en empty string, "")

'rugs'

the axes to add rugs to. Takes same values as jitter

'trend'

a vector containing the trend lines to add to the plot. Possible values are c("linear", "quadratic", "cubic")

'smooth'

the smoothing (lowess) for the points. Takes a value between 0 and 1 (the default, 0, draws no smoother)

'smoothby.lty'

the line type used for smoothers if trend.by = TRUE

'quant.smooth'

if quantile smoothers are desired, they can be specified here as either the quantiles to smooth over (e.g., c(0.25, 0.5, 0.75)), or "default", which uses the sample size to decide on an approprite set of quantile smoothers

'LOE'

logical, if TRUE, then a 1-1 line of equality is drawn

'join'

logical, if TRUE, then points are joined by lines

'lines.by

logical, if join = TRUE and colby is specified, points are joined by the specified variable

'col.trend'

a named list of colors to be used for drawing the lines. The default is list(linear = "blue", quadratic = "red", cubic = "green4")

'lty.trend'

a named list of line types for various types of trend lines. The default is list(linear = 1, quadratic = 2, cubic = 3)

'trend.by'

logical, if TRUE, then trend lines are drawn separately for each group specified by colby

'trend.parallel'

logical, if TRUE, the trend lines by group are given the same slope; otherwise they are fit independently

'col.smooth'

the colour of the smoother

'col.LOE'

the colour of the line of equality

'lty.LOE'

the line type of the line of equality

'boxplot'

logical, if TRUE, a boxplot is drawn with dotplots and histgrams

'box.lwd', 'box.col', 'box.fill'

the line width, colour, and fill colour for the box plot drawn

'bar.lwd', 'bar.col', 'bar.fill'

the line width, colour, and fill colour of bars in a bar plot

'full.height'

may no longer be necessary ...

'inf.lwd.comp', 'inf.lwd.conf'

the line width of comparison and confidence intervals, respectively

'inf.col.comp', 'inf.col.conf'

the colour of comparison and confidence intervals, respectively. These take a length 2 vector, where the first element is used for normal inference, while the second is used for bootstrap intervals

'inference.type'

the type of inference added to the plot. Possible values are c("comp", "conf")

'inference.par'

the parameter which we obtain intervals for. For a dotplot or histogram, this can be either "mean" or "median"; for bar plots it can be "proportion"

'bs.inference'

logical, if TRUE, then nonparametric bootstrap simulation is used to obtain the intervals

'min.count'

the min count for barplots inference; counts less than this are ignored

'n.boot'

the number of bootstrap simulations to perform

'large.sample.size'

sample sizes over this value will use a large-sample plot variant (i.e., scatter plots will become hex plots, dot plots become histograms)

'largesample'

logical, if TRUE, then the large-sample plot variance is used

'scatter.grid.bins'

the number, N, of bins to use for the scatter-grid plot, producing an N x N matrix

'hex.bins'

the number of bins to use for hexagonal binning

'hex.style'

the style of the hexagons, one of "size" or "alpha"

'quant.cutoff'

if quant.smooth = "default", these sample size values are used to determine which quantiles are drawn

'plottype'

used to override the default plot type. Possible values, depending on data type, include c("scatter"|"grid"|"hex"|"dot"|"hist")

'matchplots'

logical, if TRUE, then the type of plot is kept consistent between different subsets

'match.limits'

a vector of two values used to decide whether to use all small-sample or all large-sample plots

'xlim'

a vector defining the x axis limits (default NULL will use the data)

'ylim'

a vector defining the y axis limits (default NULL will use the data)

'plot.features'

a list containing any additional features for new plots (e.g., maptype)

Value

an object of class inzpar.list

Author(s)

tell029

[Package iNZightPlots version 2.7.5 Index]